Rajant IPs for VM Simulation
This commit modifies the Terraform files which create the Virtual
Machines for the Basestation, UGV1, and UAV1 (more robots will be
added later). It modifies the Virtual Network used in Azure to reflect
the network IPs of our Rajant Nodes. Rajant does not specify any
IP addresses for the gateway or VPN, so those were chosen arbitrarily.